staging/lustre: debug: quiet noisy console error messages
Quiet a number of overly noisy and unhelpful console error
messages.  Improve the format of other nearby errors.

In the case of {lod,lov}_fix_desc_stripe_size(), this doesn't
even need a console message unless it is actually changing
some stripe size that is below the minimum.  Typically it is
only zero and is being bumped up to the default value.

Lustre-commit: da430ce6173149b4bafcad71ea1d650d54ea39c0
Reviewed-on: http://review.whamcloud.com/6264
Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Bobi Jam <bobijam@gmail.com>
Reviewed-by: Brian Behlendorf <behlendorf1@llnl.gov>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>

[updated for upstream kernel submission]
Signed-off-by: Peng Tao <tao.peng@emc.com>