staging/lustre: changelog: fix CL_LAYOUT comment, string
This ports lustre commit 930fcce5978c800cc09552d8bdfffdedc3fd535a
LU-3279 changelog: fix CL_LAYOUT comment, string

Fix the CL_LAYOUT comment, since it is possible that layout swap
could be used in cases where the file content is changed.  I don't
think there are any real world use cases for this, but that depends
on what tools are being used on the filesystem.

Also change the "LAYOUT" string for this ChangeLog record type to
be "LYOUT" to match the convention of other strings to only be
5 characters long.

Intel-bug-id: LU-3279
Signed-off-by: Peng Tao <tao.peng@emc.com>