staging/lustre: fix build error on non-x86 platforms
dump_trace() is only available on X86. Without it, Lustre's own
watchdog is broken. We can only dump current task's stack.

As Oleg put it:
"
I think client-side this code is much less important and it's
probably ok to drop this altogether, since we hardly have any
ptlrpc threads on clients, most notable ones are ldlm cb threads
that should not really be blocking on the client anyway
(right, right?)

This would allow us to better prepare for the eventual update of
the kernel watchdog.
"

Remove libcfs watchdog for now. I will work on pushing it upstream
in a geneirc form.

Cc: Oleg Drokin <oleg.drokin@intel.com>
Signed-off-by: Peng Tao <tao.peng@emc.com>